+++ test/payments_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/uqpay/uqpay-sdk-go/v2/payment"
+	"github.com/uqpay/uqpay-sdk-go/v3/payment"
 )
 
 // ============================================================================